+++ src/main/java/nextstep/member/Member.java
@@ -1,6 +1,6 @@
 package nextstep.member;
 
-import auth.UserDetails;
+import auth.dto.UserDetails;
 
 public class Member {
     private Long id;
