+++ src/main/java/nextstep/member/MemberDao.java
@@ -1,7 +1,7 @@
 package nextstep.member;
 
 import auth.AuthDao;
-import auth.UserDetails;
+import auth.dto.UserDetails;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
