+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -1,10 +1,10 @@
 package nextstep.reservation;
 
-import auth.AuthenticationException;
 import auth.LoginMember;
-import auth.UserDetails;
+import auth.dto.UserDetails;
 import nextstep.member.Member;
-import org.springframework.http.HttpStatus;
+import nextstep.reservation.dto.ReservationRequest;
+import nextstep.reservation.dto.ReservationResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
