+++ src/main/java/nextstep/theme/ThemeService.java
@@ -1,5 +1,6 @@
 package nextstep.theme;
 
+import nextstep.theme.dto.ThemeRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
