+++ src/test/java/nextstep/AbstractE2ETest.java
@@ -1,9 +1,9 @@
 package nextstep;
 
 import io.restassured.RestAssured;
-import auth.TokenRequest;
-import auth.TokenResponse;
-import nextstep.member.MemberRequest;
+import auth.dto.TokenRequest;
+import auth.dto.TokenResponse;
+import nextstep.member.dto.MemberRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
